Split SignInUp into invite and SignInUp
