+++ packages/twenty-front/src/pages/auth/Invite.tsx
@@ -1,0 +1,3 @@
+import { InviteForm } from '@/auth/sign-in-up/components/InviteForm';
+
+export const Invite = () => <InviteForm />;
